chore(theme): move tokens into src/style/ and tests into test/unit/
- tokens/ -> src/style/tokens/ so all style configuration lives together
- ThemeProvider/engine.v2/verifier tests -> test/unit/... to match the
  project's test layout
